Add the entity type column (self reference back to entity catalog types).
Tests for user - belongidness of the items created, but code not there yet.. Rather, just having the capability of having a type defined.

Can items have > 1 type? What if we wanted that? At this point, it is just one.

I think that this has to become a link to a type, rather than a type. Next commit will have links and link types.